pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "winrm-mcp-server"
version = "0.1.0"
description = "MCP server for Windows Remote Management (WinRM)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.0.0",
  "pywinrm>=0.4.3",
]

[project.scripts]
winrm-mcp-server = "winrm_mcp_server.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

